+++ packages/gator-permissions-controller/src/GatorPermissionsController.test.ts
@@ -1,5 +1,5 @@
 import type { AccountSigner } from '@metamask/7715-permission-types';
-import { Messenger, deriveStateFromMetadata } from '@metamask/base-controller';
+import { deriveStateFromMetadata } from '@metamask/base-controller/next';
 import {
   createTimestampTerms,
   createNativeTokenStreamingTerms,
@@ -9,6 +9,13 @@ import {
   CHAIN_ID,
   DELEGATOR_CONTRACTS,
 } from '@metamask/delegation-deployments';
+import {
+  MOCK_ANY_NAMESPACE,
+  Messenger,
+  type MessengerActions,
+  type MessengerEvents,
+  type MockAnyNamespace,
+} from '@metamask/messenger';
 import type { HandleSnapRequest, HasSnap } from '@metamask/snaps-controllers';
 import type { SnapId } from '@metamask/snaps-sdk';
 import { hexToBigInt, numberToHex, type Hex } from '@metamask/utils';
@@ -30,10 +37,6 @@ import type {
   StoredGatorPermission,
   PermissionTypesWithCustom,
 } from './types';
-import type {
-  ExtractAvailableAction,
-  ExtractAvailableEvent,
-} from '../../base-controller/tests/helpers';
 
 const MOCK_CHAIN_ID_1: Hex = '0xaa36a7';
 const MOCK_CHAIN_ID_2: Hex = '0x1';
@@ -83,7 +86,7 @@ describe('GatorPermissionsController', () => {
   describe('constructor', () => {
     it('creates GatorPermissionsController with default state', () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
 
       expect(controller.state.isGatorPermissionsEnabled).toBe(false);
@@ -113,7 +116,7 @@ describe('GatorPermissionsController', () => {
       };
 
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
         state: customState,
       });
 
@@ -128,7 +131,7 @@ describe('GatorPermissionsController', () => {
 
     it('creates GatorPermissionsController with default config', () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
 
       expect(controller.permissionsProviderSnapId).toBe(
@@ -140,7 +143,7 @@ describe('GatorPermissionsController', () => {
 
     it('isFetchingGatorPermissions is false on initialization', () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
         state: {
           isFetchingGatorPermissions: true,
         },
@@ -153,7 +156,7 @@ describe('GatorPermissionsController', () => {
   describe('disableGatorPermissions', () => {
     it('disables gator permissions successfully', async () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
 
       await controller.enableGatorPermissions();
@@ -177,7 +180,7 @@ describe('GatorPermissionsController', () => {
   describe('fetchAndUpdateGatorPermissions', () => {
     it('fetches and updates gator permissions successfully', async () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
 
       await controller.enableGatorPermissions();
@@ -224,7 +227,7 @@ describe('GatorPermissionsController', () => {
 
     it('throws error when gator permissions are not enabled', async () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
 
       await controller.disableGatorPermissions();
@@ -240,7 +243,7 @@ describe('GatorPermissionsController', () => {
       });
 
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(rootMessenger),
+        messenger: getGatorPermissionsControllerMessenger(rootMessenger),
       });
 
       await controller.enableGatorPermissions();
@@ -262,7 +265,7 @@ describe('GatorPermissionsController', () => {
       });
 
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(rootMessenger),
+        messenger: getGatorPermissionsControllerMessenger(rootMessenger),
       });
 
       await controller.enableGatorPermissions();
@@ -286,7 +289,7 @@ describe('GatorPermissionsController', () => {
       });
 
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(rootMessenger),
+        messenger: getGatorPermissionsControllerMessenger(rootMessenger),
       });
 
       await controller.enableGatorPermissions();
@@ -302,7 +305,7 @@ describe('GatorPermissionsController', () => {
   describe('gatorPermissionsMap getter tests', () => {
     it('returns parsed gator permissions map', () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
 
       const { gatorPermissionsMap } = controller;
@@ -340,7 +343,7 @@ describe('GatorPermissionsController', () => {
       };
 
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
         state: {
           gatorPermissionsMapSerialized: JSON.stringify(mockState),
         },
@@ -354,7 +357,7 @@ describe('GatorPermissionsController', () => {
 
   describe('message handlers tests', () => {
     it('registers all message handlers', () => {
-      const messenger = getMessenger();
+      const messenger = getGatorPermissionsControllerMessenger();
       const mockRegisterActionHandler = jest.spyOn(
         messenger,
         'registerActionHandler',
@@ -382,7 +385,7 @@ describe('GatorPermissionsController', () => {
   describe('enableGatorPermissions', () => {
     it('enables gator permissions successfully', async () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
 
       await controller.enableGatorPermissions();
@@ -394,21 +397,21 @@ describe('GatorPermissionsController', () => {
   describe('metadata', () => {
     it('includes expected state in debug snapshots', () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
 
       expect(
         deriveStateFromMetadata(
           controller.state,
           controller.metadata,
-          'anonymous',
+          'includeInDebugSnapshot',
         ),
       ).toMatchInlineSnapshot(`Object {}`);
     });
 
     it('includes expected state in state logs', () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
 
       expect(
@@ -429,7 +432,7 @@ describe('GatorPermissionsController', () => {
 
     it('persists expected state', () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
 
       expect(
@@ -448,7 +451,7 @@ describe('GatorPermissionsController', () => {
 
     it('exposes expected state to UI', () => {
       const controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
 
       expect(
@@ -484,7 +487,7 @@ describe('GatorPermissionsController', () => {
 
     beforeEach(() => {
       controller = new GatorPermissionsController({
-        messenger: getMessenger(),
+        messenger: getGatorPermissionsControllerMessenger(),
       });
     });
 
@@ -693,15 +696,23 @@ describe('GatorPermissionsController', () => {
 /**
  * The union of actions that the root messenger allows.
  */
-type RootAction = ExtractAvailableAction<GatorPermissionsControllerMessenger>;
+type AllGatorPermissionsControllerActions =
+  MessengerActions<GatorPermissionsControllerMessenger>;
 
 /**
  * The union of events that the root messenger allows.
  */
-type RootEvent = ExtractAvailableEvent<GatorPermissionsControllerMessenger>;
+type AllGatorPermissionsControllerEvents =
+  MessengerEvents<GatorPermissionsControllerMessenger>;
+
+type RootMessenger = Messenger<
+  MockAnyNamespace,
+  AllGatorPermissionsControllerActions,
+  AllGatorPermissionsControllerEvents
+>;
 
 /**
- * Constructs the unrestricted messenger. This can be used to call actions and
+ * Constructs the root messenger. This can be used to call actions and
  * publish events within the tests for this controller.
  *
  * @param args - The arguments to this function.
@@ -725,8 +736,10 @@ function getRootMessenger({
 }: {
   snapControllerHandleRequestActionHandler?: HandleSnapRequest['handler'];
   snapControllerHasActionHandler?: HasSnap['handler'];
-} = {}): Messenger<RootAction, RootEvent> {
-  const rootMessenger = new Messenger<RootAction, RootEvent>();
+} = {}): RootMessenger {
+  const rootMessenger: RootMessenger = new Messenger({
+    namespace: MOCK_ANY_NAMESPACE,
+  });
 
   rootMessenger.registerActionHandler(
     'SnapController:handleRequest',
@@ -740,18 +753,27 @@ function getRootMessenger({
 }
 
 /**
- * Constructs the messenger which is restricted to relevant SampleGasPricesController
+ * Constructs the messenger supporting relevant SampleGasPricesController
  * actions and events.
  *
  * @param rootMessenger - The root messenger to restrict.
- * @returns The restricted messenger.
+ * @returns The controller messenger.
  */
-function getMessenger(
+function getGatorPermissionsControllerMessenger(
   rootMessenger = getRootMessenger(),
 ): GatorPermissionsControllerMessenger {
-  return rootMessenger.getRestricted({
-    name: 'GatorPermissionsController',
-    allowedActions: ['SnapController:handleRequest', 'SnapController:has'],
-    allowedEvents: [],
+  const gatorPermissionsControllerMessenger = new Messenger<
+    'GatorPermissionsController',
+    AllGatorPermissionsControllerActions,
+    AllGatorPermissionsControllerEvents,
+    RootMessenger
+  >({
+    namespace: 'GatorPermissionsController',
+    parent: rootMessenger,
   });
+  rootMessenger.delegate({
+    messenger: gatorPermissionsControllerMessenger,
+    actions: ['SnapController:handleRequest', 'SnapController:has'],
+  });
+  return gatorPermissionsControllerMessenger;
 }